Extract tools/linter_lib/custom_check.py.
As part of extracting this, we exempt the library from all custom
checks on itself.  This is expedient, since a lot of our
custom checks are naive about whether things are in strings, and
it is also a pain to configure individual rules.

(imported from zulip/zulip commit 40a2ffad98ab6cd0cbb6f5a8225bf01f28aba8cc)