tools/lint: Fix JSON lint pattern for spaces after ':'.
Previously, a string ending in "... :smile:" was reported as an
error and the linter complained that there should be a space
after the last ':'. This commit changes the pattern so that the
linter only checks for colons that are preceded by an opening
double-quote (").

(imported from zulip/zulip commit 10570d508f79917571b1deaa2d1f66dca7385835)